macOS compatibility

+++ play.sh
@@ -17,19 +17,30 @@
 
 MUSIC="$1"
 
-PIPE=$(mktemp -u)
-mkfifo $PIPE
-exec 3<>$PIPE
-rm $PIPE
+OS=$(uname -s)
 
-clickhouse-local --format RowBinary --query "SELECT number FROM system.numbers" >&3 &
+if [ "${OS}" = "Linux" ]
+then
+    PIPE=$(mktemp -u)
+    mkfifo $PIPE
+    exec 3<>$PIPE
+    rm $PIPE
 
-while true
-do
-    clickhouse-local --allow_experimental_analyzer 1 --format RowBinary --structure "number UInt64" --query "$(cat $MUSIC)" <&3 | aplay -f cd &
-    PID=$!
-    inotifywait -e modify $MUSIC
-    echo "Music changed." >&2
-    kill -9 $PID
-    kill -9 $(pidof aplay) 2>/dev/null
-done
+    clickhouse-local --format RowBinary --query "SELECT number FROM system.numbers" >&3 &
+
+    while true
+    do
+        clickhouse-local --allow_experimental_analyzer 1 --format RowBinary --structure "number UInt64" --query "$(cat $MUSIC)" <&3 | aplay -f cd &
+        PID=$!
+        inotifywait -e modify $MUSIC
+        echo "Music changed." >&2
+        kill -9 $PID
+        kill -9 $(pidof aplay) 2>/dev/null
+    done
+else
+    # Fallback to a simple option:
+
+    clickhouse-local --format RowBinary --query "SELECT number FROM system.numbers" | \
+        clickhouse-local --allow_experimental_analyzer 1 --format RowBinary --structure "number UInt64" --query "$(cat $MUSIC)" | \
+        play -t raw -b 16 -e signed -c 2 -v .75 -r 44100 -
+fi
